move test to it's own file so it runs in separate process, to isolate Runtime
